Shader uniform/attribute identified on creation.
When a shader is created, all uniform/samplers and attributes are
defined. The sampler/texture material methods were also redefined to for
increased flexibility. Textures can now be shared between samplers
(TODO: once two samplers are bound to the same texture unit, they cannot
be separated). Type safety was also implemented meaning that texture is
guaranteed to be the correct type (2D/Cube).