Separated lights into different structure.
Each light has a separate structure. This improves memory packing as
there is less padding within the uniform buffer. Branching in the shader
is also reduces branching as the light type does not have to be deduced.

Deferred renderering has now been modified to use multiple passes (for
each light).